TesseractPDFParser/requirements.txt
pytesseract
pdf2image
Pillow
pytest
